* Converted BagOStuff.php from the style of memcached-client.php to the standard MediaWiki style, including camel case, using protected visibility instead of initial underscore, abstract functions instead of stubs, stylize.php. * In SqlBagOStuff, ignore errors due to a read-only database, per my comments on CR r42796. Same for LocalisationCache. * Merged SqlBagOStuff and MediaWikiBagOStuff, that proved to be an awkward and unnecessary generalisation. Use the standard quoting wrapper functions instead of $db->query(). * Implemented atomic incr() and decr() functions for SqlBagOStuff. * Made incr() and decr() generally work roughly the same as it does in memcached, respecting negative steps instead of ignoring such operations. This allows decr() to be implemented in terms of incr(). * Per bug 11533, in MessageCache.php, don't retry 20 times on a cache failure, that's really memcached-specific and won't be useful for other cache types. It's not really very useful for memcached either. * Moved MySQL-specific implementations of wasDeadlock() and wasErrorReissuable() to DatabaseMysql. * Briefly tested page views with $wgReadOnly=read_only=1, fixed an error from Article::viewUpdates(). A CentralAuth fix will be in a subsequent commit.

+++ includes/LocalisationCache.php
@@ -723,6 +723,7 @@ class LCStore_DB implements LCStore {
 	var $currentLang;
 	var $writesDone = false;
 	var $dbw, $batch;
+	var $readOnly = false;
 
 	public function get( $code, $key ) {
 		if ( $this->writesDone ) {
@@ -740,17 +741,34 @@ class LCStore_DB implements LCStore {
 	}
 
 	public function startWrite( $code ) {
+		if ( $this->readOnly ) {
+			return;
+		}
 		if ( !$code ) {
 			throw new MWException( __METHOD__.": Invalid language \"$code\"" );
 		}
 		$this->dbw = wfGetDB( DB_MASTER );
-		$this->dbw->begin();
-		$this->dbw->delete( 'l10n_cache', array( 'lc_lang' => $code ), __METHOD__ );
+		try {
+			$this->dbw->begin();
+			$this->dbw->delete( 'l10n_cache', array( 'lc_lang' => $code ), __METHOD__ );
+		} catch ( DBQueryError $e ) {
+			if ( $this->dbw->wasReadOnlyError() ) {
+				$this->readOnly = true;
+				$this->dbw->rollback();
+				$this->dbw->ignoreErrors( false );
+				return;
+			} else {
+				throw $e;
+			}
+		}
 		$this->currentLang = $code;
 		$this->batch = array();
 	}
 
 	public function finishWrite() {
+		if ( $this->readOnly ) {
+			return;
+		}
 		if ( $this->batch ) {
 			$this->dbw->insert( 'l10n_cache', $this->batch, __METHOD__ );
 		}
@@ -762,6 +780,9 @@ class LCStore_DB implements LCStore {
 	}
 
 	public function set( $key, $value ) {
+		if ( $this->readOnly ) {
+			return;
+		}
 		if ( is_null( $this->currentLang ) ) {
 			throw new MWException( __CLASS__.': must call startWrite() before calling set()' );
 		}
